# test-worker: also log the pure processing time for each test case