--- a/test-projects/test-expected-error/.gitignore
+++ b/test-projects/test-expected-error/.gitignore
@@ -1,0 +1,2 @@
+rendition*
+!test/**/rendition*