--- a/test-projects/test-expected-error/manifest.yml
+++ b/test-projects/test-expected-error/manifest.yml
@@ -1,0 +1,9 @@
+packages:
+  __APP_PACKAGE__:
+    actions:
+      worker:
+        function: 'worker.js'
+        runtime: 'nodejs:10'
+        web: true
+        annotations:
+          require-adobe-auth: true