--- a/test-projects/test-expected-error/package.json
+++ b/test-projects/test-expected-error/package.json
@@ -1,0 +1,8 @@
+{
+    "name": "worker",
+    "version": "1.0.0",
+    "dependencies": {
+        "@adobe/asset-compute-sdk": "^2.6.0"
+    },
+    "private": true
+}