--- a/test-projects/test-expected-error/test/asset-compute/worker/testcase/params.json
+++ b/test-projects/test-expected-error/test/asset-compute/worker/testcase/params.json
@@ -1,0 +1,3 @@
+{
+    "errorReason": "SourceCorruptError"
+}