--- a/test-projects/test-expected-error/worker.js
+++ b/test-projects/test-expected-error/worker.js
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2019 Adobe Inc. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+'use strict'
+
+const { worker, SourceCorruptError } = require('@adobe/asset-compute-sdk');
+
+exports.main = worker(async (source, rendition) => {
+    throw new SourceCorruptError("simulated error");
+});