--- a/test-projects/test-expected-error/worker.js
+++ b/test-projects/test-expected-error/worker.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Adobe Inc. All rights reserved.
+ * Copyright 2020 Adobe Inc. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0