… test-projects/test-expected-error/test/asset-compute/worker/testcase/file.jpg