Bugfix (no jira issue)

The `test-worker` test case execution times printed in the main output include the validation script execution time. This can be confusing if someone is looking for performance data, and the validation is actually quite slow. As an example, I had a case where I used a `docker run` in the validation script which made it slower, going from `~500ms` to `1.8 seconds`.

Since the total test case execution time is still relevant to identify slow tests, this PR simply logs both. But only for successful test cases, as in case of errors the processing is "broken" anyway (?), _and_ we already log long lines with error messages in those cases.

Before:

```
Test cases:

 - testA
      ✔  Succeeded. 0s 212ms
 - testcase
      ✔  Succeeded (expected error).
 - fails
      ✖  Failure: Rendition 'rendition0.jpg' not as expected. Validate exit code was: 2. Check build/test-results/test-worker/test.log. 0s 198ms
```

After:

```
Test cases:

 - testA
      ✔  Succeeded. 0s 212ms (worker 0s 187ms)
 - testcase
      ✔  Succeeded (expected error). 0s 186ms (worker 0s 180ms)
 - fails
      ✖  Failure: Rendition 'rendition0.jpg' not as expected. Validate exit code was: 2. Check build/test-results/test-worker/test.log. 0s 198ms
```

Screenshot with colors:

<img width="453" alt="Bildschirmfoto 2020-09-29 um Di , 29  Sept , 16 13 18" src="https://user-images.githubusercontent.com/229022/94626129-478d4c80-026f-11eb-8b79-152c65596611.png">

This PR also adds a test case for "expected errors" (which wasn't covered yet) since it changes the log message for that case as well.