@alexkli 2020-09-29
@tmathern @jdelbick FYI, I made a small improvement using "worker" instead of "processing:" (shorter and IMO more intuitive to understand), as well as adjusted the color. See updated description above with screenshot.